Update config for WebPack 4
WebPack 4 requires that you specify `{ type: "javascript/auto" }` in the loader config. If you don't, the output of messageformat-loader will be passed to the default JSON loader, which will fail because the content is no longer JSON.

I'm actually not sure if it's possible to do this inside of the `require` statement.